SymPy constants fixed + Continuous classes files refactored
